Usage.md correction
I attempted to use selectize.js with a `create: function(input, callback) {}` pattern, as suggested by the docs. This wouldn't work, as the function never returns anything.. After some light debugging, I realized that the `callback` function wasn't being used or defined. I tried setting up the function given in the example, and it worked. I checked some of the examples with this definition, and they all followed the given format.