Fixed SSL issues using relative links
This switches all links which can safely be switched to a relative
protocol, to a relative protocol or to HTTPS.

GitHub Pages supports full SSL on github.io, so a few links that
targetted that have been switched to use HTTPS by default.

The GitHub buttons are (now) hosted on CloudFlare, so SSL is
available here.

While Yandex provides a SSL version of highlight.js, it doesn't
appear to be official.  The source has been switched to the cdnjs
version, as that does officially provide a SSL version and is what
is also being used for es5-shim.

The links to the external projects were not updated (that's a
different commit), and the external APIs were not touched as none
of them supported HTTPS at the time of this commit. Of course,
corsproxy is gone so that explains why the states example hasn't
worked for a while.

This closes https://github.com/brianreavis/selectize.js/issues/777.