Update utils.js
Will solve #809. Instead of checking for special keys, it just checks if the value actual has changed, It also need a bit wider space by default, to allow for two characters to be entered before growing. I have been running with this a while, and see no issues for solution, but have not tested it across all browsers.