fix(Makefile): don't replace all versions in package-lock.json during release
This change doesn't fully fix the original issue but it reduces impact area as now we only replace old version. The better solution is to replace only the first occurence or replace a version only if it is the next line after a package name (this can be done wih multiline regexps).