Fix the two closed issues #1012 and #1401
- #1401
This was still an issue at least when the options are covering multiple
lines.

- #1012
selectize "greedily" opens its select box in the multi-select case when
the options are updated. This can be seen here. This is triggered around
line 1613 in selectize.js, function addItem():

```
self.refreshOptions(self.isFocused && inputMode !== 'single');
```

refreshOptions() causes the drop-down menu to open if called with
"true". The hack-around is to fake "not-focussed" in the plugin.

Of course, the for me irritating behaviour might be a design decision.
But I do not like the drop-down to open when I just want to rearrange
the tags.