fix(a11y): blur when tabbing out of input
Fixes #1926.
To be accessible, it should be possible to easily navigate elements
using the keyboard.
Previously, when using TAB to navigate through form elements, the
input would enter its focused state when focused with TAB (and the
dropdown would open when using `openOnFocus`).
However, when TAB is used to jump to the next focusable element, it
would not lose its focused state, and the dropdown would stay open.
This commit restores the behavior before #1813 and ensures that the
input leaves its focused state and closes the dropdown when blurred
using TAB.
This commit also fixes some tests to ensure they are self-contained.